askbench/requirements.txt
openai>=1.0.0
pyyaml>=6.0
accelerate>=1.0.0
datasets>=2.20.0
peft>=0.12.0
tensorboard>=2.14.0
transformers>=4.46.0
trl>=0.23.0
